adding sonar integration - and readding sonar integration - and respective fix reported by sonar scannerspective fix reported by sonar scanner
