+++ covidAlertService/src/main/java/com/greenlearner/covidalertservice/dto/BaseDataClass.java
@@ -1,0 +1,44 @@
+package com.greenlearner.covidalertservice.dto;
+
+/**
+ * @author - GreenLearner(https://www.youtube.com/c/greenlearner)
+ */
+public class BaseDataClass {
+
+    private int confirmedCasesIndian;
+    private int confirmedCasesForeign;
+    private int discharged;
+    private int deaths;
+
+    public int getConfirmedCasesIndian() {
+        return confirmedCasesIndian;
+    }
+
+    public void setConfirmedCasesIndian(int confirmedCasesIndian) {
+        this.confirmedCasesIndian = confirmedCasesIndian;
+    }
+
+    public int getConfirmedCasesForeign() {
+        return confirmedCasesForeign;
+    }
+
+    public void setConfirmedCasesForeign(int confirmedCasesForeign) {
+        this.confirmedCasesForeign = confirmedCasesForeign;
+    }
+
+    public int getDischarged() {
+        return discharged;
+    }
+
+    public void setDischarged(int discharged) {
+        this.discharged = discharged;
+    }
+
+    public int getDeaths() {
+        return deaths;
+    }
+
+    public void setDeaths(int deaths) {
+        this.deaths = deaths;
+    }
+}
